requirements.txt
Django==3.0.4
django-bootstrap4==1.1.1
django-webpack-loader==0.7.0
djangorestframework=3.11.0